+++ src/plone/restapi/tests/test_atcontent_deserializer.py
@@ -156,7 +156,7 @@ class TestATContentDeserializer(unittest.TestCase, OrderingMixin):
 
     def test_set_layout(self):
         current_layout = self.doc1.getLayout()
-        self.assertNotEquals(current_layout, "my_new_layout")
+        self.assertNotEqual(current_layout, "my_new_layout")
         self.deserialize(body='{"layout": "my_new_layout"}')
         self.assertEqual('my_new_layout', self.doc1.getLayout())