+++ src/plone/restapi/tests/test_services.py
@@ -33,6 +33,9 @@ class TestTraversal(unittest.TestCase):
         self.api_session.headers.update({'Accept': 'application/json'})
         self.api_session.auth = (SITE_OWNER_NAME, SITE_OWNER_PASSWORD)
 
+    def tearDown(self):
+        self.api_session.close()
+
     def test_get_document(self):
         self.portal.invokeFactory(
             'Document',