+++ src/plone/restapi/tests/test_services_types.py
@@ -26,6 +26,9 @@ class TestServicesTypes(unittest.TestCase):
         self.api_session.headers.update({'Accept': 'application/json'})
         self.api_session.auth = (SITE_OWNER_NAME, SITE_OWNER_PASSWORD)
 
+    def tearDown(self):
+        self.api_session.close()
+
     def test_get_types(self):
         response = self.api_session.get(
             '{}/@types'.format(self.portal.absolute_url())
@@ -189,6 +192,9 @@ class TestServicesTypesTranslatedTitles(unittest.TestCase):
         self.api_session.auth = (SITE_OWNER_NAME, SITE_OWNER_PASSWORD)
 
         transaction.commit()
+
+    def tearDown(self):
+        self.api_session.close()
 
     def test_get_types_translated(self):
         response = self.api_session.get(